backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
python-dotenv==1.0.1
asgiref==3.8.1
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
django-steamauth==1.1.2.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
stripe==9.3.0
django-filter==24.2